requirements.txt
ollama 
streamit
pandas
numpy
datasets
transformers
sqlalchemy
sqlalchemy-redshift
langchain                 
langchain-community      
langchain-core            
langchain-openai         
langchain-pinecone        
langchainhub
openai
pinecone-client          
pinecone-text
python-dotenv
cohere
langgraph